Checkstyled and Javadocced the server and Client package

+++ src/main/java/client/Action.java
@@ -1,9 +1,10 @@
 package client;
 
 /**
- * This class adds an Action to the Database.
+ * This class represents an Action on the client side.
  */
 public class Action {
+
     private String token;
     private String action;
     private int value;
@@ -13,16 +14,15 @@ public class Action {
     private String description;
 
     /**
-     * empty constructor
+     * Empty constructor for Action.
      */
     public Action() {
     }
 
     /**
-     * constructor containing only carbon values
-     *
+     * Constructor for Action containing only the carbon values.
      * @param carbonProduced kilograms of carbon produced
-     * @param carbonReduced  kilograms of carbon produced
+     * @param carbonReduced  kilograms of carbon reduced
      */
     public Action(double carbonProduced, double carbonReduced) {
         this.carbonProduced = carbonProduced;
@@ -30,8 +30,7 @@ public class Action {
     }
 
     /**
-     * Constructs and initialises a new instance of Action.
-     *
+     * Constructor for Action containing token, action name and the amount of points.
      * @param token  the token
      * @param action the action name
      * @param value  the amount of points
@@ -43,16 +42,16 @@ public class Action {
     }
 
     /**
-     * Create a new action.
-     *
+     * Constructor for Action containing token, action name, the amount of points,
+     * the amount of carbon produced and the amount of carbon reduced.
      * @param token          the token
      * @param action         the action
      * @param value          the value
      * @param carbonProduced carbon produced
-     * @param carbonReduced  carbon saced
+     * @param carbonReduced  carbon reduced
      */
-    public Action(String token, String action, int value, double carbonReduced, double carbonProduced) {
-
+    public Action(String token, String action, int value,
+                  double carbonReduced, double carbonProduced) {
         this.token = token;
         this.action = action;
         this.value = value;
@@ -61,16 +60,17 @@ public class Action {
     }
 
     /**
-     * Create a new action.
-     *
+     * Constructor for Action containing token, action name, the amount of points,
+     * the amount of carbon produced, the amount of carbon reduced and the date of the action.
      * @param token          the token
      * @param action         the action
      * @param value          the value
      * @param carbonProduced carbon produced
-     * @param carbonReduced  carbon saced
-     * @param date           date added (in miliseconds)
+     * @param carbonReduced  carbon reduced
+     * @param date           date added (in milliseconds)
      */
-    public Action(String token, String action, int value, int carbonReduced, int carbonProduced, long date) {
+    public Action(String token, String action, int value, double carbonReduced,
+                  double carbonProduced, long date) {
         this.token = token;
         this.action = action;
         this.value = value;
@@ -80,8 +80,8 @@ public class Action {
     }
 
     /**
-     * Create a new action.
-     *
+     * Constructor for Action containing token, action name, the amount of points,
+     * the amount of carbon produced and the description.
      * @param token          the token
      * @param action         the action
      * @param value          the value
@@ -99,15 +99,16 @@ public class Action {
     }
 
     /**
-     * Create a new action.
-     *
+     * Constructor for Action containing action name, the amount of points,
+     * the amount of carbon produced, the amount of carbon reduced and the date of the action.
      * @param action         the action
      * @param value          the value
      * @param carbonProduced carbon produced
-     * @param carbonReduced  carbon saced
-     * @param date           date added (in miliseconds)
+     * @param carbonReduced  carbon reduced
+     * @param date           date added (in milliseconds)
      */
-    public Action(String action, int value, int carbonReduced, int carbonProduced, long date) {
+    public Action(String action, int value, double carbonReduced, double carbonProduced,
+                  long date) {
         this.action = action;
         this.value = value;
         this.carbonReduced = carbonReduced;
@@ -117,7 +118,6 @@ public class Action {
 
     /**
      * getter for token.
-     *
      * @return token
      */
     public String getToken() {
@@ -126,6 +126,7 @@ public class Action {
 
     /**
      * setter for token.
+     * @param token token
      */
     public void setToken(String token) {
         this.token = token;
@@ -133,7 +134,6 @@ public class Action {
 
     /**
      * getter for Action.
-     *
      * @return action
      */
     public String getAction() {
@@ -142,6 +142,7 @@ public class Action {
 
     /**
      * setter for Action.
+     * @param action action name
      */
     public void setAction(String action) {
         this.action = action;
@@ -149,7 +150,6 @@ public class Action {
 
     /**
      * getter for Value.
-     *
      * @return value
      */
     public int getValue() {
@@ -158,6 +158,7 @@ public class Action {
 
     /**
      * setter for Value.
+     * @param value value
      */
     public void setValue(int value) {
         this.value = value;
@@ -165,7 +166,6 @@ public class Action {
 
     /**
      * getter for carbonReduced.
-     *
      * @return carbonReduced
      */
     public double getCarbonReduced() {
@@ -174,14 +174,14 @@ public class Action {
 
     /**
      * setter for carbonReduced.
+     * @param carbonReduced carbonReduced
      */
     public void setCarbonReduced(double carbonReduced) {
         this.carbonReduced = carbonReduced;
     }
 
     /**
-     * getter for Value.
-     *
+     * getter for carbonProduced.
      * @return carbonProduced
      */
     public double getCarbonProduced() {
@@ -189,39 +189,40 @@ public class Action {
     }
 
     /**
-     * setter for Value.
+     * setter for carbonProduced.
+     * @param carbonProduced carbon Produced
      */
     public void setCarbonProduced(double carbonProduced) {
         this.carbonProduced = carbonProduced;
     }
 
     /**
-     * setter for date.
+     * getter for date.
+     * @return date
      */
     public long getDate() {
         return date;
     }
 
     /**
-     * getter for date.
-     *
-     * @return date
+     * setter for date.
+     * @param date date
      */
     public void setDate(long date) {
         this.date = date;
     }
 
     /**
-     * setter for description.
+     * getter for description.
+     * @return description
      */
     public String getDescription() {
         return description;
     }
 
     /**
-     * getter for description.
-     *
-     * @return description
+     * setter for description.
+     * @param description description
      */
     public void setDescription(String description) {
         this.description = description;
